requirements.txt
PyQt5==5.15.11
PyOpenGL==3.1.7
numpy
pyqtgraph==0.13.7